Add files via upload
This is a c++ program that solves the problem of a semi- Hanoi by using a Hanoi function.
In semi-Hanoi, we assume that the discs are numbered. Then we try to move the discs so as we end up with odd discs on one tower and even on the other. The rule of Hanoi tower i.e. a disc placed above another disc should be of smaller size applies here also.